pyproject.toml
[project]
name = "electronic-survey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "requests>=2.32",
    "semantic-version>=2.10",
]
